Refactor: Modularize core components and update documentation
- I reorganized the codebase into distinct modules:
  - Matrix library moved to `src/math/`
  - NeuroNet and NeuroNetLayer moved to `src/neural_network/`
  - GeneticAlgorithm moved to `src/optimization/`
- I changed the namespace for GeneticAlgorithm to `Optimization`.
- I updated include paths in all affected C++ files.
- I updated `src/Neuro Net DLL.vcxproj` and `src/Neuro Net DLL.vcxproj.filters` to reflect new file locations for `neuronet.h` and `neuronet.cpp`.
- I identified that a `ProjectReference` in `src/Neuro Net DLL.vcxproj` to a missing `../includes/Neuro Net Includes.vcxproj` file is likely a legacy Windows configuration and not relevant for the current CMake-based Linux build. I did not modify this file.
- I updated `README.md` to reflect the new directory structure, component locations, and updated include paths in code examples.
- I created `CHANGELOG.md` to document these changes.
