+++ src/optimization/genetic_algorithm.cpp
@@ -14,7 +14,7 @@
 #include <limits>    // For std::numeric_limits
 #include <stdexcept> // For std::runtime_error (optional, for error handling)
 
-namespace NeuroNet {
+namespace Optimization {
 
 /**
  * @brief Constructs a GeneticAlgorithm instance.
@@ -23,7 +23,7 @@ namespace NeuroNet {
  * The template_network is copied and used as a blueprint for all individuals
  * in the population, ensuring consistent network architecture.
  */
-GeneticAlgorithm::GeneticAlgorithm(
+Optimization::GeneticAlgorithm::GeneticAlgorithm(
     int population_size,
     double mutation_rate,
     double crossover_rate,
@@ -54,7 +54,7 @@ GeneticAlgorithm::GeneticAlgorithm(
  * values, typically between -1.0 and 1.0.
  * @return NeuroNet A new, randomly initialized NeuroNet individual.
  */
-NeuroNet GeneticAlgorithm::create_random_individual() const {
+NeuroNet Optimization::GeneticAlgorithm::create_random_individual() const {
     NeuroNet individual = template_network_; // Start with the template structure.
 
     // Retrieve the total number of weights and biases from the template structure.
@@ -87,7 +87,7 @@ NeuroNet GeneticAlgorithm::create_random_individual() const {
  * new individuals, each created by `create_random_individual()`.
  * Fitness scores are also cleared, as they will need to be re-evaluated.
  */
-void GeneticAlgorithm::initialize_population() {
+void Optimization::GeneticAlgorithm::initialize_population() {
     population_.clear();
     fitness_scores_.clear(); // Fitness scores are now invalid.
     population_.reserve(population_size_); // Pre-allocate memory.
@@ -112,7 +112,7 @@ void GeneticAlgorithm::initialize_population() {
  * but does not update the overall `best_individual_` or `best_fitness_score_` (across all generations);
  * that update is handled in `evolve_one_generation` after selection and other operations.
  */
-void GeneticAlgorithm::evaluate_fitness(const std::function<double(NeuroNet&)>& fitness_function) {
+void Optimization::GeneticAlgorithm::evaluate_fitness(const std::function<double(NeuroNet&)>& fitness_function) {
     if (population_.empty()) {
         // Or throw std::runtime_error("Population is empty, cannot evaluate fitness.");
         return;
@@ -142,7 +142,7 @@ void GeneticAlgorithm::evaluate_fitness(const std::function<double(NeuroNet&)>& 
  * @return const NeuroNet& A reference to the selected parent individual.
  * @throws std::runtime_error if the population is empty.
  */
-const NeuroNet& GeneticAlgorithm::tournament_selection(int tournament_size) const {
+const NeuroNet& Optimization::GeneticAlgorithm::tournament_selection(int tournament_size) const {
     if (population_.empty()) {
         throw std::runtime_error("Tournament selection called on an empty population.");
     }
@@ -182,7 +182,7 @@ const NeuroNet& GeneticAlgorithm::tournament_selection(int tournament_size) cons
  *    c. Otherwise (no crossover), the selected parents are mutated and added.
  * The `population_` member is updated with the new generation.
  */
-void GeneticAlgorithm::selection() {
+void Optimization::GeneticAlgorithm::selection() {
     if (population_.empty() || fitness_scores_.empty() || fitness_scores_.size() != population_.size()) {
         // Or throw std::runtime_error("Population or fitness scores are not ready for selection.");
         return;
@@ -246,7 +246,7 @@ void GeneticAlgorithm::selection() {
  * @param parent2 The second parent NeuroNet.
  * @return std::vector<NeuroNet> A vector containing two new offspring NeuroNet individuals.
  */
-std::vector<NeuroNet> GeneticAlgorithm::crossover(const NeuroNet& parent1, const NeuroNet& parent2) {
+std::vector<NeuroNet> Optimization::GeneticAlgorithm::crossover(const NeuroNet& parent1, const NeuroNet& parent2) {
     NeuroNet offspring1 = template_network_; // Ensure offspring have the correct structure.
     NeuroNet offspring2 = template_network_;
 
@@ -305,7 +305,7 @@ std::vector<NeuroNet> GeneticAlgorithm::crossover(const NeuroNet& parent1, const
  * by a small random value (e.g., adding a value between -0.1 and 0.1).
  * @param individual The NeuroNet to mutate (modified in place).
  */
-void GeneticAlgorithm::mutate(NeuroNet& individual) {
+void Optimization::GeneticAlgorithm::mutate(NeuroNet& individual) {
     std::uniform_real_distribution<double> prob_dist(0.0, 1.0); // For checking against mutation_rate_
     std::uniform_real_distribution<float> mutation_val_dist(-0.1f, 0.1f); // Magnitude of mutation
 
@@ -341,7 +341,7 @@ void GeneticAlgorithm::mutate(NeuroNet& individual) {
  * 3. Performing selection (which internally handles crossover and mutation for the new population).
  * @param fitness_function The function used to evaluate the fitness of each individual.
  */
-void GeneticAlgorithm::evolve_one_generation(const std::function<double(NeuroNet&)>& fitness_function) {
+void Optimization::GeneticAlgorithm::evolve_one_generation(const std::function<double(NeuroNet&)>& fitness_function) {
     if (population_.empty()) {
         // This might happen if initialize_population was not called before the first evolution step.
         // Or if population_size_ is 0.
@@ -393,7 +393,7 @@ void GeneticAlgorithm::evolve_one_generation(const std::function<double(NeuroNet
  * It first initializes the population, then iteratively calls `evolve_one_generation`.
  * @param fitness_function The function to evaluate individual fitness.
  */
-void GeneticAlgorithm::run_evolution(const std::function<double(NeuroNet&)>& fitness_function) {
+void Optimization::GeneticAlgorithm::run_evolution(const std::function<double(NeuroNet&)>& fitness_function) {
     initialize_population(); // Prepare the initial random population.
 
     for (int i = 0; i < num_generations_; ++i) {
@@ -412,7 +412,7 @@ void GeneticAlgorithm::run_evolution(const std::function<double(NeuroNet&)>& fit
  * (which could be default if no better one was found).
  * @return NeuroNet A copy of the best performing individual.
  */
-NeuroNet GeneticAlgorithm::get_best_individual() const {
+NeuroNet Optimization::GeneticAlgorithm::get_best_individual() const {
     // If best_individual_ was never updated (e.g. evolution didn't run or no valid individuals),
     // it might be default-constructed.
     // A check could be added: if best_fitness_score_ is still numeric_limits::lowest(),
@@ -432,4 +432,4 @@ NeuroNet GeneticAlgorithm::get_best_individual() const {
     return best_individual_; // Return the best found across generations
 }
 
-} // namespace NeuroNet
+} // namespace Optimization
